AI/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
scikit-learn>=1.3.0
numpy>=1.26.0
anthropic>=0.25.0
openai>=1.30.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
